requirements.txt
appdirs==1.4.4
asgiref==3.3.2
black==20.8b1
click==7.1.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.5.0
et-xmlfile==1.0.1
isort==5.8.0
MarkupPy==1.14
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.7
pathspec==0.8.1
Pillow==8.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
xlrd==2.0.1
xlwt==1.3.0
